requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.1
httpx>=0.26.0
python-docx>=1.1.0
pypdf>=4.0.0
python-multipart>=0.0.9
Pillow>=10.2.0
easyocr>=1.7.1
opencv-python-headless>=4.9.0
google-generativeai>=0.3.2
surya-ocr>=0.4.0
google-generativeai>=0.3.2
